requirements.txt
boto3==1.4.0
botocore==1.4.58
dj-database-url==0.4.1
Django==1.10
django-storages==1.5.0
docutils==0.12
gunicorn==19.6.0
jmespath==0.9.0
Pillow==3.3.1
psycopg2==2.6.2
python-dateutil==2.5.3
pytz==2016.6.1
s3transfer==0.1.5
six==1.10.0
sqlparse==0.2.1
whitenoise==3.2.1
djangorestframework==3.4.6